+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # 0.0.1
 
 - initial release
+
+# 2.0.0
+
+- uses a canvas with a raw image instead of a material (easier lighting)
+- hides background objects
+- added tooltips
+- fixed unnecessary console logs in Edit mode
+- added a way to override your animator